Refactor download command [WEB-1540]
The download command was running into memory leak errors because
it used to `json_decode` then `json_encode` the giant block of
data that the Shop API returns. This meant we could save the JSON
file locally formatted for readability, but it was a taxing process
on data we're already getting in the format we need. This commit
removed the prettifying logic so that the command works in the
memory environments of our servers.